labs/lab04-graphrag-gremlin/src/requirements.txt
gremlinpython>=3.7.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-mgmt-search>=9.0.0
azure-mgmt-cognitiveservices>=13.0.0
azure-mgmt-cosmosdb>=9.0.0
openai>=1.10.0
python-dotenv>=1.0.0
